Move ServerMain to nomadrealms.app and update pom
Relocate ServerMain into the nomadrealms.app package (file moved/renamed and package declaration updated). Update pom.xml mainClass to nomadrealms.app.ServerMain so the build points to the new entry point.
