+++ nomad-realms-server/src/main/java/nomadrealms/app/ServerMain.java
@@ -1,4 +1,4 @@
-package nomadrealms;
+package nomadrealms.app;
 
 import engine.nengen.Nengen;
 import nomadrealms.app.context.ServerContext;
