+++ nomad-realms-server/src/main/java/nomadrealms/networking/HostedWorld.java
@@ -1,0 +1,7 @@
+package nomadrealms.networking;
+
+/**
+ * R
+ */
+public class HostedWorld {
+}
